+++ src/idom/backend/_urls.py
@@ -1,0 +1,7 @@
+from pathlib import PurePosixPath
+
+
+PATH_PREFIX = PurePosixPath("/_idom")
+MODULES_PATH = PATH_PREFIX / "modules"
+ASSETS_PATH = PATH_PREFIX / "assets"
+STREAM_PATH = PATH_PREFIX / "stream"