+++ src/idom/backend/default.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import asyncio
-from typing import Any
+from typing import Any, NoReturn
 
 from idom.types import RootComponentConstructor
 
@@ -21,6 +21,11 @@ def configure(
 def create_development_app() -> Any:
     """Create an application instance for development purposes"""
     return _default_implementation().create_development_app()
+
+
+def Options(*args: Any, **kwargs: Any) -> NoReturn:
+    """Create configuration options"""
+    raise ValueError("Default implementation has no options.")  # pragma: no cover
 
 
 async def serve_development_app(