+++ tests/test_backend/test_utils.py
@@ -1,4 +1,3 @@
-import asyncio
 import threading
 import time
 from contextlib import ExitStack
@@ -11,7 +10,6 @@ from idom.backend.utils import find_available_port
 from idom.backend.utils import run as sync_run
 from idom.backend.utils import traversal_safe_path
 from idom.sample import SampleApp as SampleApp
-from tests.tooling.loop import open_event_loop
 
 
 @pytest.fixture
@@ -28,22 +26,19 @@ def test_find_available_port():
 
 
 async def test_run(page: Page, exit_stack: ExitStack):
-    loop = exit_stack.enter_context(open_event_loop(as_current=False))
-
     host = "127.0.0.1"
     port = find_available_port(host)
     url = f"http://{host}:{port}"
 
-    def run_in_thread():
-        asyncio.set_event_loop(loop)
-        sync_run(
+    threading.Thread(
+        target=lambda: sync_run(
             SampleApp,
             host,
             port,
             implementation=flask_implementation,
-        )
-
-    threading.Thread(target=run_in_thread, daemon=True).start()
+        ),
+        daemon=True,
+    ).start()
 
     # give the server a moment to start
     time.sleep(0.5)